chore(EMS-3842): rename function
