+++ src/api/helpers/map-CIS-countries/map-CIS-country/index.test.ts
@@ -6,7 +6,7 @@ import canGetAQuoteOnline from './can-get-a-quote-online';
 import canGetAQuoteByEmail from './can-get-a-quote-by-email';
 import cannotGetAQuote from './cannot-get-a-quote';
 import canApplyForInsuranceOnline from './can-apply-for-insurance-online';
-import canApplyOffline from './can-apply-offline';
+import canApplyOffline from './can-apply-for-quote-offline';
 import noInsuranceSupportAvailable from './no-insurance-support';
 import { EXTERNAL_API_DEFINITIONS, EXTERNAL_API_MAPPINGS } from '../../../constants';
 import { MappedCisCountry } from '../../../types';
